Add null check for request.File in Create method
Updated the Create method in ItemService to include a null check for `request.File` before uploading. This prevents null reference exceptions by ensuring the `file` variable is only assigned if a file is provided. The assignment of `ImageId` has also been modified to use the null-conditional operator (`?.`) for safer access to the `Id` property.